refactor: rename varint_slop
