+++ packages/vexide-startup/src/varint_encoding.rs
